keyboard: prepare for high-byte-set keycodes
Using a uint16_t scancode will allow more leeway in the codes we emit without being constrained by ASCII

+++ kernel/include/drivers/ps2/keyboard.h
@@ -33,7 +33,7 @@ class PS2Keyboard : public PS2Controller::Device {
             keymap_to_use keymap;
             bool ctrldown;
             bool altdown;
-            uint8_t keycode;
+            uint16_t keycode;
             bool down;
 
             key_event_t();
